Project: Classify
This commit breaks outs many of the existing classes into their
own files.  This allows for more people to work on the project
at a time and for re-use when needed.
